Adds Vic's office state management (#2091)
* Updates lower hallway occupancy entity ID

* Update office state manager entity ID

* Adds mushroom lamp to presence detection

Extends the office presence detection automation to include the mushroom lamp.

This ensures the mushroom lamp is turned on when presence is detected
and turned off after a timeout when no presence is detected. The state
of the mushroom lamp is also tracked in the state manager.

* Adds Vic's office state management

Implements a state management system for Vic's office to persist the desired state of devices (lights, fan) even when the office is unoccupied.

This ensures that when the office becomes occupied again, devices return to their previously desired state.

It introduces a variable entity to store the "should be" states and automations to manage device states based on occupancy. Includes timeout to reset the 'should be' state after a period if the device is turned off.

---------

Co-authored-by: Will Garside <worgarside@gmail.com>