Add 5s delay for kitchen light dimming (#2130)
* Add 5s delay for kitchen light dimming

* Improves basement presence detection reliability

* Changes automation mode to queued

Updates the automation to use queued mode with a maximum of 2 concurrent executions.

This prevents potential race conditions and ensures that automations are processed in the order they are triggered, up to a limit of 2 simultaneous runs.

* Changes automation mode to restart

Updates the automation mode from queued to restart.

This change ensures that the automation restarts whenever the trigger conditions are met, providing more responsive behavior. The previous 'queued' mode could lead to delays if multiple triggers occurred in quick succession.

* Rounds brightness percentages to nearest integer

Refactors brightness calculations to round the calculated percentage to the nearest whole number using the 'ceil' method. This ensures that the light brightness is always set to a valid integer value, preventing potential errors.

* Delays light activation in the kitchen.

Adds a short delay before turning on the kitchen light.

This prevents immediate light activation when presence is
briefly detected, leading to a less intrusive experience.

* Corrects presence detection during evening hours

Adjusts the time condition for presence detection to accurately
activate between 12:00 and 23:00, ensuring proper functionality
during the intended evening hours.

---------

Co-authored-by: Will Garside <worgarside@gmail.com>