Add hot water switch (#2374)
* Add hot water switch

* Add hot water automations

* Adds proximity-based hot water control

Implements automation to turn on the hot water when someone approaches home, but only when the schedule is active and the hot water is not already on.

Introduces an input number to configure the proximity distance threshold.

* Turns off hot water based on proximity

Adds an automation that turns off the hot water when a person moves beyond a threshold distance and is traveling away from home.

This helps conserve energy by only heating water when needed.

---------

Co-authored-by: Will Garside <worgarside@gmail.com>