Improves boiler control logic (#2403)
* Improves boiler control logic

* Refactors boiler switch logic

Simplifies the switch logic for central heating and hot water.

Now uses the boiler LED sensor to determine the state and conditions for turning on/off the switches, creating a more robust solution.

* Updates sensor state management

Uses `value_template` instead of `state` for more accurate sensor readings.

This ensures the sensor state is derived from the template, providing reliable feedback from the boiler.

* Refactors switch icon logic

Updates the switch icon templates to directly reflect the state
of the relevant binary sensor, ensuring accurate representation.

This improves the reliability of the heating and hot water switch
icons by binding their status to the actual boiler state.

---------

Co-authored-by: Will Garside <worgarside@gmail.com>
Co-authored-by: Will Garside <26073454+worgarside@users.noreply.github.com>